Loading launches by SWR and refactor into components
