+++ components/templates/index.ts
@@ -5,12 +5,10 @@ export * from "./showDate"
 export * from "./showResolution"
 export * from "./showStatusIcon"
 export * from "./showTestStats"
-export * from "./showAndChangeLaunchStatus"
 export * from "./TestExpandablePanels"
 export * from "./TestResolution"
 export * from "./ListOfSuites"
 export * from "./ListOfTests"
 export * from "./TestHistory"
-export * from "./DeltaViewForLaunches"
 export * from "./eventNotification"
 export * from "./DeleteProject"
